Check Variable sizes
Sizes of all linked Variables are now checked, by checking that
the size of the ReactionMethod Domain for a VariableReaction matches
the size of the VariableDomain that it is linked to.

It is possible to opt-out of this check by setting the VariableAttribute
:check_length=false. This is needed for a few cases including:

- a ReactionMethod in an interior Domain calculates boundary fluxes assuming
  a regular column grid, and needs to add them to a surface or floor
  Variable: the surface/floor Variable will need the :check_length=false
  attribute, as it is not the same size as the interior Domain.

- a setup ReactionMethod configures all grid variables, including
  both interior and boundary Variables (which will need the :check_length=false
  attribute).

If it is necessary to use the :check_length=false attribute, then
any loop that accesses these Variables should not use @inbounds so that
the low-level Julia checks will still catch configuration errors.