Implement copy(VariableReaction) instead of using deepcopy
This is an optimisation to reducing startup time by speeding up
create reaction methods.

NB: only a shallow copy of the attributes Dict is made. This should be safe
as the attribute values are currently all immutable: will need reviewing if attributes
are ever used in the future for mutable objects that shouldn't be shared.