ModelData can now store multiple sets of arrays for different element types
Work in progress (PALEOboxes tests pass, but code is incomplete)

This simplifies managing data arrays when eg using ForwardDiff for automatic
differentiation, which requires an additional set of data arrays with Dual number types.

Arrays in the set are identified by arrays_idx, which now needs to be supplied
along with modeldata for all data access methods.
The default arrays (usually eltype Float64) have arrays_idx=1

Additional array types can be added to modeldata with

    add_arrays_eltype!

and removed with

   pop_arrays_data!

allocate_variables! is modified so that data arrays that should be shared with
base Float64 arrays (constant values, or data that is not calculated by a partial
Jacobian etc) are now shared, instead of requiring ad hoc copying.

dispatch_setup! is modified so that :setup only updates the base array set
(arrays_idx=1), and these values are then copied to other arrays in modeldata.