Define cartesian_size for all grids
Define 'cartesian_size' for all grids, including 'UnstructuredVectorGrid' and 'UnstructuredColumnGrid',
returning a single dimension with length = number of cells.

This allows generic code eg in ReactionForceGrid to work for these grids, assuming the supplied forcing
file has a single spatial dimension corresponding to the total number of cells in the model grid.