Add new check_variable_links method
Add
    check_variable_links(model, modeldata; [throw_on_error=true] [, expect_hostdep_varnames=["global.tforce"]]) -> links_ok::Bool

This checks that all Variables are linked correctly, by checking that there are no unexpected host-dependent non-state
Variables (ie unlinked Variables).

This should be called soon after 'create_model_from_config', so that an appropriate error is generated early in the user model script
if any variables are unlinked.

Remove the variable link check from 'check_ready', as that was too late to be useful (often a hard-to-interpret error to do with
Variable size or allocation had already happened).

See corresponding PR in PALEOmodel to add this check to PALEOmodel.initialize!