Better error message if a ReactionMethod fails at runtime
Catch runtime exception if a ReactionMethod fails, and print an
error message containing tje name of the failed Reaction as in the YAML
config file. This is helpful if there are multiple instances of the same Reaction
and it's not clear which one is failing.

NB: requires 'generated_dispatch=false' argument to PALEOmodel.initialize!