Better error message for misconfiguration with no _sms variable
Better error message for a rarely seen misconfiguration, where the
<X>_sms derivative variable corresponding to state variable <X> is either missing,
or has a Target (not just Contributors) defined.

The only way I can see to create a configuration like this would be to accidentally configure a
FluxTarget reaction that includes the <X>_sms variable, or a as a result of a programming error
in a .jl file defining a Reaction.